resource_availability/requirements.txt
cloudshell-automation-api>=7.1,<7.2
pyyaml==3.12
